requirements.txt
# Core dependencies
langgraph>=0.2.0
langchain-core>=0.3.0
langchain-anthropic>=0.2.0
anthropic>=0.39.0
openai>=1.0.0
mcp>=1.1.0
pydantic>=2.0.0
python-dotenv>=1.0.0
fastapi>=0.109.0
uvicorn>=0.27.0

# Development dependencies
pytest>=8.0.0
black>=24.0.0
ruff>=0.7.0
mypy>=1.0.0
